tests.bash: trap the signals SIGINT
The bash script tests.bash exits with exit code 0 even if it is
signaled.

The script should exit with the number of the signal when it is
signaled; that is possible if the signal is trapped.

This commit traps the signal SIGINT that causes the bash script
tests.bash to exit in error if it is interrupted (^C).