tests.bash: fix test against the variable ko
The variable ko is left unset if no tests end in error; its value is
empty in that situation; it cannot be an integer though. If set, its
value is at least 1 and thus cannot be 0.

This commit fixes the test against the variable ko by testing if it is
set.